Moved everything into a flit based package
