--- etca_asm/version
@@ -1,1 +1,0 @@
-0.0.2
