+++ src/etc_as/__init__.py
@@ -1,0 +1,5 @@
+"""
+A configurable assembler for ETCa
+"""
+
+__version__ = open(__file__.rstrip('__init__.py') + 'version').read()
