+++ src/etc_as/__main__.py
@@ -1,0 +1,7 @@
+from __future__ import annotations
+
+import sys
+
+from etc_as.main import main
+
+sys.exit(main())
