+++ src/etc_as/extensions/__init__.py
@@ -2,6 +2,6 @@ import pkgutil
 
 
 def import_all_extensions():
-    import etca_asm.base_isa
+    import etc_as.base_isa
     for args in pkgutil.iter_modules(__path__):
         __import__(f'{__name__}.{args.name}')
